separate internal v external utils
