+++ packages/progress-bar/src/testing/index.ts
@@ -1,2 +1,1 @@
 export * from './getTestUtils';
-export * from './utils';
